requirements.txt
# ----------------------
# Flask
# ----------------------
flask
flask-SQLALchemy
psycopg2
Flask-WTF


# ----------------------
# Testing
# ----------------------
mock


# ----------------------
# Production Server
# ----------------------
gunicorn
